move queries code into transform
